chore: rename eslint-rule-tester to test-helpers
