+++ package.json
@@ -117,6 +117,6 @@
         "*.md": "markdownlint --fix"
     },
     "imports": {
-        "#eslint-rule-tester": "./tests/eslint-rule-tester.js"
+        "#test-helpers": "./tests/test-helpers.js"
     }
 }
