+++ tests/lib/rules/global-require.js
@@ -4,7 +4,7 @@
  */
 "use strict"
 
-const RuleTester = require("#eslint-rule-tester").RuleTester
+const RuleTester = require("#test-helpers").RuleTester
 const rule = require("../../../lib/rules/global-require")
 
 const ERROR = { messageId: "unexpected", type: "CallExpression" }
