+++ tests/lib/rules/no-hide-core-modules.js
@@ -14,7 +14,7 @@
 //------------------------------------------------------------------------------
 
 const path = require("path")
-const RuleTester = require("#eslint-rule-tester").RuleTester
+const RuleTester = require("#test-helpers").RuleTester
 const rule = require("../../../lib/rules/no-hide-core-modules")
 
 //------------------------------------------------------------------------------
