+++ tests/lib/rules/no-missing-require.js
@@ -5,7 +5,7 @@
 "use strict"
 
 const path = require("path")
-const { RuleTester, isCaseSensitiveFileSystem } = require("#eslint-rule-tester")
+const { RuleTester, isCaseSensitiveFileSystem } = require("../../test-helpers")
 const rule = require("../../../lib/rules/no-missing-require")
 
 const tsReactExtensionMap = [
