+++ tests/lib/rules/no-new-require.js
@@ -4,7 +4,7 @@
  */
 "use strict"
 
-const RuleTester = require("#eslint-rule-tester").RuleTester
+const RuleTester = require("#test-helpers").RuleTester
 const rule = require("../../../lib/rules/no-new-require")
 
 new RuleTester().run("no-new-require", rule, {
