+++ tests/lib/rules/no-path-concat.js
@@ -5,7 +5,7 @@
 "use strict"
 
 const path = require("path")
-const RuleTester = require("#eslint-rule-tester").RuleTester
+const RuleTester = require("#test-helpers").RuleTester
 const rule = require("../../../lib/rules/no-path-concat")
 
 new RuleTester().run("no-path-concat", rule, {
