+++ tests/lib/rules/no-process-env.js
@@ -4,7 +4,7 @@
  */
 "use strict"
 
-const RuleTester = require("#eslint-rule-tester").RuleTester
+const RuleTester = require("#test-helpers").RuleTester
 const rule = require("../../../lib/rules/no-process-env")
 
 new RuleTester().run("no-process-env", rule, {
