+++ tests/lib/rules/no-restricted-require.js
@@ -5,7 +5,7 @@
 "use strict"
 
 const path = require("path")
-const RuleTester = require("#eslint-rule-tester").RuleTester
+const RuleTester = require("#test-helpers").RuleTester
 const rule = require("../../../lib/rules/no-restricted-require")
 
 new RuleTester().run("no-restricted-require", rule, {
