+++ tests/lib/rules/no-sync.js
@@ -4,7 +4,7 @@
  */
 "use strict"
 
-const RuleTester = require("#eslint-rule-tester").RuleTester
+const RuleTester = require("#test-helpers").RuleTester
 const rule = require("../../../lib/rules/no-sync")
 
 new RuleTester().run("no-sync", rule, {
